Rework g2.mec dim
g2.mec's dim method is now able to get 2 new properties.
off for an offset of the given coordinates. this makes it more
convenient to use it to show the dimension lines for 2 points.
over is a new property to define the overshooting of the lines which
indicates the offset.
g2.mec's dim can now accept styling parameters and uses the vec method
for its arrows. This makes them easier to maintain.