Update img and remove optional styles from docs
The img parameters are changed from:
`uri, x, y ,b ,h , sx, sy, sb ,sh , xoff, yoff, w, scl`
to
`uri, x, y, b, h, xoff, yoff, dx, dy, w, scl`,
so that:

- `xoff` and `yoff` take the place of `sb` and `sh`
- `dx` and `dy` replace `sb` and `sh`

The tests are going to be updated in an upcoming change.

The API documents state that no defaults are given from `b` and `h`, but they are still set to the image width and height if undefined.

The optional parameters referring styles are removed from
the docs.